refactor: alternative 패키지 생성
